requirements.txt
pyperclip
google-generativeai
datetime
win10toast
